+++ Source/ImGui/Private/ImGuiContextProxy.cpp
@@ -83,24 +83,34 @@ FImGuiContextProxy::FImGuiContextProxy(const FString& InName, int32 InContextInd
 	BeginFrame();
 }
 
-void FImGuiContextProxy::Draw()
+void FImGuiContextProxy::DrawEarlyDebug()
 {
-	if (bIsFrameStarted && !bIsDrawCalled)
+	if (bIsFrameStarted && !bIsDrawEarlyDebugCalled)
 	{
-		bIsDrawCalled = true;
+		bIsDrawEarlyDebugCalled = true;
 
 		SetAsCurrent();
 
-		// Broadcast draw event to allow listeners to draw their controls to this context.
-#if DRAW_EVENTS_ORDER_WORLD_BEFORE_MULTI_CONTEXT
-		BroadcastWorldTick();
-#endif // DRAW_EVENTS_ORDER_WORLD_BEFORE_MULTI_CONTEXT
+		// Delegates called in order specified in FImGuiDelegates.
+		BroadcastMultiContextEarlyDebug();
+		BroadcastWorldEarlyDebug();
+	}
+}
 
-		BroadcastMultiContextTick();
+void FImGuiContextProxy::DrawDebug()
+{
+	if (bIsFrameStarted && !bIsDrawDebugCalled)
+	{
+		bIsDrawDebugCalled = true;
 
-#if !DRAW_EVENTS_ORDER_WORLD_BEFORE_MULTI_CONTEXT
-		BroadcastWorldTick();
-#endif // !DRAW_EVENTS_ORDER_WORLD_BEFORE_MULTI_CONTEXT
+		// Make sure that early debug is always called first to guarantee order specified in FImGuiDelegates.
+		DrawEarlyDebug();
+
+		SetAsCurrent();
+
+		// Delegates called in order specified in FImGuiDelegates.
+		BroadcastWorldDebug();
+		BroadcastMultiContextDebug();
 	}
 }
 
@@ -116,7 +126,7 @@ void FImGuiContextProxy::Tick(float DeltaSeconds)
 		if (bIsFrameStarted)
 		{
 			// Make sure that draw events are called before the end of the frame.
-			Draw();
+			DrawDebug();
 
 			// Ending frame will produce render output that we capture and store for later use. This also puts context to
 			// state in which it does not allow to draw controls, so we want to immediately start a new frame.
@@ -149,7 +159,8 @@ void FImGuiContextProxy::BeginFrame(float DeltaTime)
 		ImGui::NewFrame();
 
 		bIsFrameStarted = true;
-		bIsDrawCalled = false;
+		bIsDrawEarlyDebugCalled = false;
+		bIsDrawDebugCalled = false;
 	}
 }
 
@@ -186,7 +197,28 @@ void FImGuiContextProxy::UpdateDrawData(ImDrawData* DrawData)
 	}
 }
 
-void FImGuiContextProxy::BroadcastWorldTick()
+void FImGuiContextProxy::BroadcastWorldEarlyDebug()
+{
+	if (ContextIndex != Utilities::INVALID_CONTEXT_INDEX)
+	{
+		FSimpleMulticastDelegate& WorldEarlyDebugEvent = FImGuiDelegatesContainer::Get().OnWorldEarlyDebug(ContextIndex);
+		if (WorldEarlyDebugEvent.IsBound())
+		{
+			WorldEarlyDebugEvent.Broadcast();
+		}
+	}
+}
+
+void FImGuiContextProxy::BroadcastMultiContextEarlyDebug()
+{
+	FSimpleMulticastDelegate& MultiContextEarlyDebugEvent = FImGuiDelegatesContainer::Get().OnMultiContextEarlyDebug();
+	if (MultiContextEarlyDebugEvent.IsBound())
+	{
+		MultiContextEarlyDebugEvent.Broadcast();
+	}
+}
+
+void FImGuiContextProxy::BroadcastWorldDebug()
 {
 	if (DrawEvent.IsBound())
 	{
@@ -203,7 +235,7 @@ void FImGuiContextProxy::BroadcastWorldTick()
 	}
 }
 
-void FImGuiContextProxy::BroadcastMultiContextTick()
+void FImGuiContextProxy::BroadcastMultiContextDebug()
 {
 	if (SharedDrawEvent && SharedDrawEvent->IsBound())
 	{