+++ Source/ImGui/Public/ImGuiDelegates.h
@@ -6,15 +6,36 @@
 
 
 /**
- * Delegates to ImGui debug events called. World delegates are called once per frame to draw debug for owning world
- * and are automatically cleared when that world becomes invalid. Multi-context delegates are called once for every
- * world that needs to be debugged.
- * In engine version 4.18 or later delegates are called during OnWorldPostActorTick event while in older versions
- * 4.18 during OnWorldTickStart event. Calling behaviours can be changed in build configuration.
+ * Delegates to ImGui debug events. World delegates are called once per frame during world updates and have invocation
+ * lists cleared after their worlds become invalid. Multi-context delegates are called once for every updated world.
+ * Early debug delegates are called during world tick start and debug delegates are called during world post actor tick
+ * or in engine versions below 4.18 during world tick start.
+ *
+ * Order of events is defined in a way that multi-context delegates can be used to draw headers and/or footers:
+ * multi-context early debug, world early debug, world debug, multi-context debug.
  */
 class IMGUI_API FImGuiDelegates
 {
 public:
+
+	/**
+	 * Get a delegate to ImGui world early debug event for current world (GWorld).
+	 * @returns Simple multicast delegate to debug events called once per frame to debug current world
+	 */
+	static FSimpleMulticastDelegate& OnWorldEarlyDebug();
+
+	/**
+	 * Get a delegate to ImGui world early debug event for given world.
+	 * @param World - World for which we need a delegate
+	 * @returns Simple multicast delegate to debug events called once per frame to debug given world
+	 */
+	static FSimpleMulticastDelegate& OnWorldEarlyDebug(UWorld* World);
+
+	/**
+	 * Get a delegate to ImGui multi-context early debug event.
+	 * @returns Simple multicast delegate to debug events called once per frame for every world to debug
+	 */
+	static FSimpleMulticastDelegate& OnMultiContextEarlyDebug();
 
 	/**
 	 * Get a delegate to ImGui world debug event for current world (GWorld).